drv: mailbox: no size_t to avoid unistd.h
The unistd.h also defines sleep(), but we want our own sleep,
so as a workaround for now just avoid the header instead
of having to rename sleep() to something else.

+++ drivers/mailbox.c
@@ -201,7 +201,7 @@ int mbox_release(struct mbox *m)
     return 0;
 }
 
-int mbox_send(struct mbox *m, uint32_t *msg, size_t len)
+int mbox_send(struct mbox *m, uint32_t *msg, unsigned len)
 {
     unsigned i;
